review ccsbase export
